Prepare for project renaming
 - refactoring with relative imports to support simple renaming
 - new tox job to build and simply test code inside docker container
