+++ Dockerfile.env.example
@@ -1,6 +1,6 @@
 LAOS_HOST=0.0.0.0
 LAOS_PORT=10001
-LAOS_DB=mysql://root:ubuntu@192.168.0.120/functions
+LAOS_DB=mysql+pymysql://root:ubuntu@192.168.0.120/functions
 KEYSTONE_ENDPOINT=http://192.168.0.120:5000/v3
 FUNCTIONS_URL=http://192.168.0.120:8080/v1
 LAOS_LOG_LEVEL=INFO
