+++ README.md
@@ -168,111 +168,6 @@ Once server is launched you can navigate to:
 
 to see recent API docs
 
-Testing (general information)
------------------------------
-
-In order to run tests you need to install `Tox`:
-
-    $ pip install tox
-
-Also, you will need a running MySQL instance with the database migrations applied from the previous step.
-Tests are dependent on pre-created MySQL database for persistence.
-Please set env var
-
-    $ export TEST_DB_URI=mysql://<your-user>:<your-user-password>@<mysql-host>:<mysql-port>/<functions-db>
-
-Testing: PEP8
--------------
-
-In order to run `PEP8` style checks run following command:
-
-    $ tox -e pep8
-
-
-Testing: Functional
--------------------
-
-In order to run `functional` tests run following command:
-
-    $ tox -e py35-functional
-
-Pros:
-
-* lightweight (controllers and DB models testing)
-* no OpenStack required
-* no IronFunctions required
-
-Cons:
-
-* MySQL server required
-* OpenStack authentication is not tested
-* IronFunctions API stubbed with fake implementation
-
-Testing: Integration
---------------------
-
-Integration tests are dependent on following env variables:
-
-* TEST_DB_URI - similar to functional tests, database endpoint
-* FUNCTIONS_API_URL - IronFunctions API URL (default value - `http://localhost:8080/v1`)
-* OS_AUTH_URL - OpenStack Identity endpoint
-* OS_PROJECT_NAME - OpenStack user-specific project name
-* OS_USERNAME - OpenStack user name
-* OS_PASSWORD - OpenStack user user password
-
-To run tests use following command:
-
-    export TEST_DB_URI=mysql://<your-user>:<your-user-password>@<mysql-host>:<mysql-port>/<functions-db>
-    export FUNCTIONS_API_URL=<functions-api-protocol>://<functions-host>:<functions-port>/<functions-api-version>
-    export OS_AUTH_URL=<identity-api-protocol>://<identity-host>:<identity-port>/<identity-api-version>
-    export OS_PROJECT_NAME=<project-name>
-    export OS_USERNAME=<project-name>
-    export OS_PASSWORD=<project-name>
-    tox -epy35-integration
-
-Testing: Coverage regression
-----------------------------
-
-In order to build quality software it is necessary to keep test coverage at its highest point.
-So, as part of `Tox` testing new check was added - functional test coverage regression.
-In order to run it use following command:
-
-    $ tox -e py35-functional-regression
-
-3rd party bugs to resolve
--------------------------
-
-IronFunctions:
-
-* https://github.com/iron-io/functions/issues/298
-* https://github.com/iron-io/functions/issues/296
-* https://github.com/iron-io/functions/issues/275
-* https://github.com/iron-io/functions/issues/274
-
-TODOs
------
-
-Swagger doc:
-
-* Make swagger doc more explicit on HTTP POST/UPDATE body content
-* HTTP headers requests
-
-IronFunctions:
-
-* Support app deletion in IronFunctions
-* Support tasks listing/showing
-
-Laos:
-
-* Tests: integration, functional, units
-* Better logging coverage
-
-Python Functions client:
-
-* Support logging instance passing in [function-python](https://github.com/iron-io/functions_python)
-* python-laosclient (ReST API client and CLI tool)
-* App writing examples
-
 
 Contacts
 --------
