+++ laos/api/controllers/routes.py
@@ -19,9 +19,9 @@ from aiohttp import web
 from aioservice.http import controller
 from aioservice.http import requests
 
-from laos.api.views import app as app_view
-from laos.common import config
-from laos.models import app as app_model
+from ...common import config
+from ...models import app as app_model
+from ..views import app as app_view
 
 
 class AppRouteV1Controller(controller.ServiceController):
