+++ laos/tests/integration/test_apps.py
@@ -12,8 +12,8 @@
 #    License for the specific language governing permissions and limitations
 #    under the License.
 
-from laos.tests.common import apps as apps_suite
-from laos.tests.integration import base
+from ..common import apps as apps_suite
+from ..integration import base
 
 
 class TestIntegrationApps(base.LaosIntegrationTestsBase,
