+++ tox.ini
@@ -1,7 +1,7 @@
 # Project LaOS
 
 [tox]
-envlist = py35-functional,py35-functional-regression,py35-integration,py35-integration-regression,pep8
+envlist = py35-functional,py35-functional-regression,py35-integration,py35-integration-regression,pep8,docker-build
 minversion = 1.6
 skipsdist = True
 
@@ -14,16 +14,16 @@ passenv =
         OS_PASSWORD
         OS_USERNAME
         OS_PROJECT_NAME
+        DOCKER_HOST
 setenv = VIRTUAL_ENV={envdir}
 usedevelop = True
 install_command = pip install -U {opts} {packages}
 deps = -r{toxinidir}/requirements.txt
        -r{toxinidir}/test-requirements.txt
 commands = find . -type f -name "*.pyc" -delete
-           rm -f .testrepository/times.dbm
-           python setup.py testr --testr-args='{posargs}'
 whitelist_externals = find
                       rm
+                      docker
 
 [testenv:pep8]
 commands = flake8
@@ -48,7 +48,10 @@ commands =
     rm -rf doc/html doc/build
     python setup.py build_sphinx
 
+[testenv:docker-full]
+commands = {toxinidir}/scripts/docker_full.sh
+
 [flake8]
-ignore = H202,H404,H405,H501
+ignore = H202,H304,H404,H405,H501
 show-source = True
 exclude=.venv,.git,.tox,dist,doc,*lib/python*,*egg,build,migrations
